+++ packages/histoire-plugin-vue2/src/client/app/MountStory.ts
@@ -12,9 +12,9 @@ import {
 } from '@histoire/vendors/vue'
 import type { Story } from '@histoire/shared'
 // @ts-expect-error virtual module id
-import * as setup from '$histoire-setup'
+import * as setup from 'virtual:$histoire-setup'
 // @ts-expect-error virtual module id
-import * as generatedSetup from '$histoire-generated-global-setup'
+import * as generatedSetup from 'virtual:$histoire-generated-global-setup'
 import { registerGlobalComponents } from './global-components.js'
 import { RouterLinkStub } from './RouterLinkStub'